Last commit message was bogus.
I was working on two projects at the same time and entered the commit
message for my other project into this one and pushed to GitHub before I
could amend it.  The last commit for this projects should have been to
the effect that I completed the description of the Java bindings.  Oh,
and I converted a readme.txt to a readme.md.